services/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
httpx==0.27.0
numpy==1.26.4
scikit-learn==1.5.0
Pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
